+++ CHANGELOG.md
@@ -145,6 +145,7 @@ These PRs are not directly user-facing, but improve the development experience. 
 - **Elements**: All hooks are now free functions (i.e. `use_state(hooks, ..)` instead of `hooks.use_state(..)`)
 - **UI**: Focus is now global across different packages, and we've removed the FocusRoot component
 - **API**: CursorLockGuard removed and `hide_cursor` package introduced.
+- **Hierarchies**: The `children` component is now automatically derived from `parent` components (unless the user opts out of this). The `children` component is also not networked any longer, since it's calculated on the client side.
 
 #### Non-breaking